// web/src/lib/openapi.d.ts
export interface components {
    schemas: {
// ...
        CreateTokenRequest: {
            label: string;
            pat: string;
            scopes?: string;
            /** Format: date-time */
            expires_at?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}